refactor: update dependencies and remove most circular deps
BREAKING CHANGE: Refactores type check functions i.e. isEnumType --> ApiBuilderEnum.isInstanceOf
