+++ package.json
@@ -5,10 +5,12 @@
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "scripts": {
+    "check-types": "tsc --noEmit",
     "build": "tsc --project ./tsconfig.build.json",
     "clean": "rm -rf dist",
     "coverage": "jest --coverage",
-    "lint": "tslint -c tslint.json -p tsconfig.json",
+    "lint": "eslint . --ext .js,.ts",
+    "lint:fix": "eslint . --ext .js,.ts --fix",
     "prepublishOnly": "run-s clean build",
     "test": "jest"
   },
@@ -29,22 +31,24 @@
   "homepage": "https://github.com/apicollective/apibuilder-js#readme",
   "dependencies": {
     "invariant": "^2.2.4",
-    "lodash": "^4.17.15",
+    "lodash": "^4.17.21",
     "pluralize": "^8.0.0"
   },
+  "sideEffects": false,
   "devDependencies": {
-    "@types/faker": "^4.1.9",
-    "@types/invariant": "^2.2.31",
-    "@types/jest": "^24.0.25",
-    "@types/lodash": "^4.14.149",
-    "@types/node": "^10.14.19",
+    "@types/faker": "^5.5.8",
+    "@types/invariant": "^2.2.35",
+    "@types/jest": "^27.0.1",
+    "@types/lodash": "^4.14.172",
+    "@types/node": "^16.9.1",
     "@types/pluralize": "0.0.29",
-    "faker": "^4.1.0",
-    "jest": "^24.9.0",
-    "npm-run-all": "^4.1.5",
-    "ts-jest": "^24.3.0",
-    "tslint": "^5.20.1",
-    "tslint-config-airbnb": "^5.11.2",
-    "typescript": "^3.7.4"
+    "@typescript-eslint/eslint-plugin": "^4.31.0",
+    "eslint": "^7.32.0",
+    "eslint-config-airbnb-base-typescript": "^1.0.0",
+    "eslint-plugin-jest": "^24.4.0",
+    "faker": "^5.5.3",
+    "jest": "^27.1.1",
+    "ts-jest": "^27.0.5",
+    "typescript": "^4.4.2"
   }
 }
