+++ src/language/index.ts
@@ -1,4 +1,4 @@
 export * from './ast';
 export * from './constants';
 export * from './FullyQualifiedName';
-export * from './predicates';
+export { default as FullyQualifiedName } from './FullyQualifiedName';
