+++ src/type/predicates.ts
@@ -1,55 +1,10 @@
-import {
-  ApiBuilderArray,
-  ApiBuilderEnclosingType,
-  ApiBuilderEnum,
-  ApiBuilderMap,
-  ApiBuilderModel,
-  ApiBuilderPrimitiveType,
-  ApiBuilderType,
-  ApiBuilderUnion,
-} from './definition';
-
-/**
- * Returns whether the specified object is an API Builder enumeration type.
- */
-export function isEnumType(type: any): type is ApiBuilderEnum {
-  return type instanceof ApiBuilderEnum;
-}
-
-/**
- * Returns whether the specified object is an API Builder array type.
- */
-export function isArrayType(type: any): type is ApiBuilderArray {
-  return type instanceof ApiBuilderArray;
-}
-
-/**
- * Returns whether the specified object is an API Builder map type.
- */
-export function isMapType(type: any): type is ApiBuilderMap {
-  return type instanceof ApiBuilderMap;
-}
-
-/**
- * Returns whether the specified object is an API Builder model type.
- */
-export function isModelType(type: any): type is ApiBuilderModel {
-  return type instanceof ApiBuilderModel;
-}
-
-/**
- * Returns whether the specified object is an API Builder primitive type.
- */
-export function isPrimitiveType(type: any): type is ApiBuilderPrimitiveType {
-  return type instanceof ApiBuilderPrimitiveType;
-}
-
-/**
- * Returns whether the specified object is an API Builder union type.
- */
-export function isUnionType(type: any): type is ApiBuilderUnion {
-  return type instanceof ApiBuilderUnion;
-}
+import type { ApiBuilderEnclosingType, ApiBuilderType } from './types';
+import ApiBuilderArray from './ApiBuilderArray';
+import ApiBuilderEnum from './ApiBuilderEnum';
+import ApiBuilderMap from './ApiBuilderMap';
+import ApiBuilderModel from './ApiBuilderModel';
+import ApiBuilderPrimitiveType from './ApiBuilderPrimitiveType';
+import ApiBuilderUnion from './ApiBuilderUnion';
 
 /**
  * Returns whether the specified object is one of the possible
@@ -57,8 +12,8 @@ export function isUnionType(type: any): type is ApiBuilderUnion {
  */
 export function isEnclosingType(type: any): type is ApiBuilderEnclosingType {
   return (
-    isArrayType(type) ||
-    isMapType(type)
+    ApiBuilderArray.isInstanceOf(type)
+    || ApiBuilderMap.isInstanceOf(type)
   );
 }
 
@@ -68,12 +23,12 @@ export function isEnclosingType(type: any): type is ApiBuilderEnclosingType {
  */
 export function isType(type: any): type is ApiBuilderType {
   return (
-    isArrayType(type) ||
-    isMapType(type) ||
-    isPrimitiveType(type) ||
-    isModelType(type) ||
-    isEnumType(type) ||
-    isUnionType(type)
+    ApiBuilderArray.isInstanceOf(type)
+    || ApiBuilderMap.isInstanceOf(type)
+    || ApiBuilderPrimitiveType.isInstanceOf(type)
+    || ApiBuilderModel.isInstanceOf(type)
+    || ApiBuilderEnum.isInstanceOf(type)
+    || ApiBuilderUnion.isInstanceOf(type)
   );
 }
 
@@ -81,7 +36,9 @@ export function isType(type: any): type is ApiBuilderType {
  * If a given type is an enclosing type, this recursively strips the enclosing
  * wrappers and returns the underlying type.
  */
-export function getBaseType(type: ApiBuilderType): Exclude<ApiBuilderType, ApiBuilderEnclosingType> {
+export function getBaseType(
+  type: ApiBuilderType,
+): Exclude<ApiBuilderType, ApiBuilderEnclosingType> {
   if (isEnclosingType(type)) {
     return getBaseType(type.ofType);
   }
