+++ test/specs/ApiBuilderImport.spec.ts
@@ -1,5 +1,11 @@
 import keyBy from 'lodash/keyBy';
-import { ApiBuilderImport, ApiBuilderService, isEnumType, isModelType, isUnionType } from '../../src';
+import {
+  ApiBuilderImport,
+  ApiBuilderService,
+  ApiBuilderEnum,
+  ApiBuilderModel,
+  ApiBuilderUnion,
+} from '../../src';
 import apidocApi from '../fixtures/apidocApi';
 
 const service = new ApiBuilderService(apidocApi);
@@ -15,38 +21,38 @@ describe('ApiBuilderImport', () => {
   });
 
   test('should have properties with types from specified schema', () => {
-    expect(subject.enums.every(isEnumType)).toBe(true);
-    expect(subject.models.every(isModelType)).toBe(true);
-    expect(subject.unions.every(isUnionType)).toBe(true);
+    expect(subject.enums.every(ApiBuilderEnum.isInstanceOf)).toBe(true);
+    expect(subject.models.every(ApiBuilderModel.isInstanceOf)).toBe(true);
+    expect(subject.unions.every(ApiBuilderUnion.isInstanceOf)).toBe(true);
   });
 });
 
 describe('ApiBuilderImport::findEnumByName', () => {
   test('should return type matching short name', () => {
-    expect(isEnumType(subject.findEnumByName('parameter_location'))).toBe(true);
+    expect(ApiBuilderEnum.isInstanceOf(subject.findEnumByName('parameter_location'))).toBe(true);
   });
 
   test('should return type matching fully qualified name', () => {
-    expect(isEnumType(subject.findEnumByName('com.bryzek.apidoc.spec.v0.enums.parameter_location'))).toBe(true);
+    expect(ApiBuilderEnum.isInstanceOf(subject.findEnumByName('com.bryzek.apidoc.spec.v0.enums.parameter_location'))).toBe(true);
   });
 });
 
 describe('ApiBuilderImport::findModelByName', () => {
   test('should return type matching short name', () => {
-    expect(isModelType(subject.findModelByName('operation'))).toBe(true);
+    expect(ApiBuilderModel.isInstanceOf(subject.findModelByName('operation'))).toBe(true);
   });
 
   test('should return type matching fully qualified name', () => {
-    expect(isModelType(subject.findModelByName('com.bryzek.apidoc.spec.v0.models.operation'))).toBe(true);
+    expect(ApiBuilderModel.isInstanceOf(subject.findModelByName('com.bryzek.apidoc.spec.v0.models.operation'))).toBe(true);
   });
 });
 
 describe('ApiBuilderImport::findUnionByName', () => {
   test('should return type matching short name', () => {
-    expect(isUnionType(subject.findUnionByName('response_code'))).toBe(true);
+    expect(ApiBuilderUnion.isInstanceOf(subject.findUnionByName('response_code'))).toBe(true);
   });
 
   test('should return type matching fully qualified name', () => {
-    expect(isUnionType(subject.findUnionByName('com.bryzek.apidoc.spec.v0.unions.response_code'))).toBe(true);
+    expect(ApiBuilderUnion.isInstanceOf(subject.findUnionByName('com.bryzek.apidoc.spec.v0.unions.response_code'))).toBe(true);
   });
 });
