+++ test/specs/ApiBuilderOperation.spec.ts
@@ -6,9 +6,9 @@ describe('ApiBuilderOperation', () => {
     test('returns type matching response code', () => {
       const service = new ApiBuilderService(apidocApi);
       const resource = service.resources
-        .find(resource => resource.typeName === 'application');
+        .find((res) => res.typeName === 'application');
       const operation = resource?.operations
-        .find(operation => operation.path === '/:orgKey' && operation.method === 'GET');
+        .find((oprtn) => oprtn.path === '/:orgKey' && oprtn.method === 'GET');
       const type = operation?.getResponseTypeByCode(200);
       expect(type?.toString()).toEqual('[com.bryzek.apidoc.api.v0.models.application]');
     });
