--- tslint.json
@@ -1,9 +1,0 @@
-{
-  "extends": "tslint-config-airbnb",
-  "rules": {
-    "max-line-length": [true, {
-      "limit": 120,
-      "ignore-pattern": "^import |^export {(.*?)}"
-    }]
-  }
-}
